+++ packages/grid-export/package.json
@@ -33,6 +33,7 @@
     "@stenajs-webui/elements": "7.3.0",
     "@stenajs-webui/grid": "7.3.0",
     "@types/zipcelx": "^1.5.0",
+    "clipboard-polyfill": "^3.0.2",
     "date-fns": "2.16.1",
     "zipcelx": "^1.6.2"
   },
